Change make clean to not remove dirs inside BINDIR
This change was discussed in #3.

Previously, make clean would recursively remove empty directories inside
BINDIR. This was too intrusive, as it could cause problems if the user
had set BINDIR to ".".
Now, only BINDIR itself is deleted (if it exists and is empty). This
change is not possible for OBJDIR and DEPDIR, since the build process
may add directory trees in them to reflect the source file directory
tree.