Reduce verbosity of compilation commands
To handle various edge cases and to support various features, the
compilation commands of the Makefile have gotten very complex. They
often even exceed the width of the terminal in length.

The full compilation command is rarely of interest. Much more important
is instantly seeing what file is being compiled. That is the motivation
of this change. In fact, we were already hiding some details, such as
POSTCOMPILE. Advanced users that want to see the full compilation command
are likely advanced enough to derive it from the Makefile.