Add some color to Makefile output using tput
Added coloring to filenames outputted by the Makefile using tput. This
makes it easier to see to which file a certain block of compiler output
belongs. And it looks nice.

The coloring can easily be configured using a new configuration
variable. The default coloring is bold yellow. This fits well with the
default colors used by gcc (bold red and bold cyan). Underlining would
be a safer default coloring, but that would make filenames with
underscores slightly harder to read (and it looks worse, in my opinion).

The Makefile tests if tput is installed, and silently omits the coloring
if it is not, so this change does not make the Makefile depend on tput.