Rearranged config variables slightly
Moved the binary names to the top and grouped CXX with the various flag
variables.

The binary names are the variables that the user will most often change:
they are the only variables that will be different for almost every
project. Therefore, they should be at the top. The newproject script
does set these variables automatically, but users may not always use
the script.

The grouping of the variables in the usage instructions is now also the
same as in the settings.