Replaced "::=" with ":="
The "::=" syntax is POSIX and ":=" is not, which is probably why we went
with "::=" before. However, GNU Make also supports ":=", and has done so
for a longer time than "::=". I have already encountered one machine
with such an old version of make that "::=" was not supported, but ":="
was. Hence this change.

If we already use any GNU Make-specific features, which we probably do,
this change only improves portability.

See also:
https://stackoverflow.com/questions/37206926/for-immediate-variable-assignment-double-colon-or-one